refactor: Implement core functionality for AutoExitNode and split into modules
- Added network management features in `network.go` to retrieve current SSID and check cellular connectivity.
- Introduced startup management in `startup.go` to create, check, and remove startup shortcuts for the application.
- Integrated Tailscale management in `tailscale.go` to activate and deactivate exit nodes based on network conditions.
- Developed system tray functionality in `tray.go` to display network status, manage startup options, and handle updates.
- Implemented update checking and notification system in `update.go` to inform users of new releases via GitHub API.
